LA-91 Rework - Move a global function to a new utils class (to avoid globals).
